commit working patch from trunk/ as starting point
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/concert-pitch-matrix@11459 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ src/gui/editors/matrix/MatrixScene.cpp
@@ -204,7 +204,7 @@ MatrixScene::getPriorSegment()
 Segment *
 MatrixScene::getNextSegment()
 {
-    if (m_currentSegmentIndex + 1 >= m_segments.size()) return 0;
+    if ((unsigned int) m_currentSegmentIndex + 1 >= m_segments.size()) return 0;
     return m_segments[m_currentSegmentIndex+1];
 }
